\name{Conifers2012}
\alias{Conifers2012}
\docType{data}
\title{
Leslie et al. 2012 conifers
}
\description{
This has 1000 BEAST chronograms for 492 conifer and related species.
}
\usage{data(Conifers2012)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Leslie, AB, Beaulieu JM, Rai HS, Crane PR, Donoghue MJ, Mathews S.  2012.  Hemisphere-scale differences in conifer evolutionary dynamics. Proceedings of the National Academy of Sciences. 109(40): 16217-16221.
}

\examples{
#not run
#data(Conifers2012)
#plot(Conifers2012[[1]],show.node.label=TRUE)
}
\keyword{datasets}
