\name{CowmanBellwood2011}
\alias{Labridae2011}
\alias{Pomacentridae2011}
\alias{Chaetodontidae2011}
\alias{Apogonidae2011}
\docType{data}
\title{
Trees of coral reef fishes
}
\description{
This has four different groups, with a chronogram for each: Labridae2011, Pomacentridae2011, Chaetodontidae2011, Apogonidae2011.
}
\usage{data(Labridae2011)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Cowman PF, Bellwood DR. 2011. Coral reefs as drivers of cladogenesis: expanding coral reefs, cryptic extinction events, and the development of biodiversity hotspots. Journal of Evolutionary Biology 24(12): 2543-2562
}

\examples{
#not run
#data(Pomacentridae2011)
#plot(Pomacentridae2011,show.node.label=TRUE)
}
\keyword{datasets}
