\name{FabreEtAl2009}
\alias{FabreEtAl2009}
\docType{data}
\title{
Fabre et al. 2009 primate tree
}
\description{
This has a single chronogram (FabreEtAl2009) from this paper. The tree is in phylo format. Note the dryad file for this paper (doi:10.5061/dryad.8492) also includes a larger tree with congeners missing from the analysis attached at the MRCA of their congeners; this tree has just the 273 taxa for which there was data.
}
\usage{data(FabreEtAl2009)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Fabre P, Rodrigues A, Douzery EJP. 2009. Patterns of macroevolution among Primates inferred from a supermatrix of mitochondrial and nuclear DNA. Molecular Phylogenetics and Evolution 53(3): 808-825
}

\examples{
#not run
#data(FabreEtAl2009)
#plot(FabreEtAl2009,show.node.label=TRUE)
}
\keyword{datasets}
