\name{HardyCook2012}
\alias{HardyCook2012}
\docType{data}
\title{
Hardy and Cook 2012 Ericaceae
}
\description{
This has one chronogram for Ericaceae.
}
\usage{data(HardyCook2012)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Hardy NB, Cook LG. 2012. Testing for ecological limitation of diversification: a case study using parasitic plants. American Naturalist 180(4): 438-449
}

\examples{
#not run
#data(HardyCook2012)
#plot(HardyCook2012,show.node.label=TRUE)
}
\keyword{datasets}
