\name{LopezFernandezEtAl2012}
\alias{LopezFernandezEtAl2012}
\docType{data}
\title{
Lopez-Fernandez et al. 2012 Neotropical cichlids
}
\description{
This has 1000 BEAST chronograms for Neotropical cichlids. They are stored in a multiphylo object.
}
\usage{data(LopezFernandezEtAl2012)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
López-Fernández H, Arbour JH, Winemiller KO, Honeycutt RL (2012) Testing for ancient adaptive radiations in Neotropical cichlid fishes. Evolution 67(5): 1321-1337
}

\examples{
#not run
#data(LopezFernandezEtAl2012)
#plot(LopezFernandezEtAl2012[[17]],show.node.label=TRUE)
}
\keyword{datasets}
