\name{LopezFernandezEtAl2013}
\alias{LopezFernandezEtAl2013}
\docType{data}
\title{
Lopez-Fernandez et al. 2013 Neotropical cichlids
}
\description{
This has 1000 BEAST chronograms for Neotropical cichlids. They are stored in a multiphylo object.
}
\usage{data(LopezFernandezEtAl2013)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Pyron RA. 2011. Divergence time estimation using fossils as terminal taxa and the origins of Lissamphibia. Systematic Biology 60(4): 466-481.
}

\examples{
#not run
#data(LopezFernandezEtAl2013)
#plot(LopezFernandezEtAl2013[[17]],show.node.label=TRUE)
}
\keyword{datasets}
