\name{NauheimerEtAl2012}
\alias{NauheimerEtAl2012}
\docType{data}
\title{
Nauheimer et al. 2012 tree of Araceae
}
\description{
This has a single timetree (NauheimerEtAl2012$trees$best) from this paper. The tree is in phylo format.
}
\usage{data(NauheimerEtAl2012)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Nauheimer L., Metzler D., & Renner S.S. 2012. Global history of the ancient monocot family Araceae inferred with models accounting for past continental positions and previous ranges inferred from fossils. New Phytologist, 195(4): 938-950.
}

\examples{
#not run
#data(NauheimerEtAl2012)
#plot(NauheimerEtAl2012$trees$best,show.node.label=TRUE)
}
\keyword{datasets}
