\name{Oaks2011}
\alias{Oaks2011}
\docType{data}
\title{
Oaks 2011 tree of crocodylians
}
\description{
This has a single chronogram (Oaks2011) from this paper. The tree is in phylo (\pkg{ape}) format.
}
\usage{data(Oaks2011)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Oaks, J. R. 2011. A time-calibrated species tree of crocodylia reveals a recent
radiation of the true crocodiles. Evolution 11: 3285-3297. 
}

\examples{
#not run
#data(Oaks2011)
#crocTree<-Oaks2011
#plot the tree
#plot(crocTree,show.node.label=TRUE)

}
\keyword{datasets}
