\name{Pyron2011}
\alias{Pyron2011}
\docType{data}
\title{
Pyron tree of Lissamphibia, including fossil taxa
}
\description{
This has two trees made by BEAST, one with a single clock (Pyron2011[[1]]), the other with two clocks (Pyron2011[[2]]). Note that some of the terminal taxa are NOT extant, so the tree is a chronogram but not ultrametric}
\usage{data(Pyron2011)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Pyron RA. 2011. Divergence time estimation using fossils as terminal taxa and the origins of Lissamphibia. Systematic Biology 60(4): 466-481.
}

\examples{
#not run
#data(Pyron2011)
#plot(Pyron2011[[1]],show.node.label=TRUE)
}
\keyword{datasets}
