\name{SantiniEtAl2009}
\alias{SantiniEtAl2009}
\docType{data}
\title{
Santini et al. tree of teleost fishes}
\description{
This has a chronogram of teleost fishes.}
\usage{data(SantiniEtAl2009)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
 F. Santini, L.J. Harmon, G. Carnevale, and M. E. Alfaro. 2009. Did genome duplication drive the origin of teleosts? A comparative study of diversification in ray-finned fishes. BMC Evolutionary Biology 9:194
}

\examples{
#not run
#data(SantiniEtAl2009)
#plot(SantiniEtAl2009,show.node.label=TRUE)
}
\keyword{datasets}
