\name{SmithEtAl2011}
\alias{SmithEtAl2011}
\docType{data}
\title{
Smith et al. 2011 angiosperm tree
}
\description{
This has the best (SmithEtAl2011_tree) likelihood topology. The tree is in phylo (\pkg{ape}) format.
}
\usage{data(SmithEtAl2011)}
\details{
Make sure to cite the paper and dryad reference below when you use this dataset.
}
\source{
Smith SA, Beaulieu JM, Stamatakis A, Donoghue MJ (2011) Understanding angiosperm diversification using small and large phylogenetic trees. American Journal of Botany, online in advance of print. doi:10.3732/ajb.1000481

Smith SA, Beaulieu JM, Stamatakis A, Donoghue MJ (2011) Data from: Understanding angiosperm diversification using small and large phylogenetic trees. Dryad Digital Repository. doi:10.5061/dryad.8790
}
\examples{
#not run
#data(SmithEtAl2011)

#library(ape)
#angTree<-SmithEtAl2011_tree

#plot the tree
#plot(angTree,show.node.label=TRUE)
}
\keyword{datasets}
