\name{Unpub1}
\alias{Unpub1}
\docType{data}
\title{
Eastman et al. unpublished tree of salamanders
}
\description{
This has a single timetree of salamanders (Unpub1). The tree is in phylo (\pkg{ape}) format.
}
\usage{data(Unpub1)}
\details{
Tree is unpublished, contact Jon Eastman for more information}
\source{

}

\examples{
#not run
#data(Unpub1)
#salamanderTree<-Unpub1
#plot(salamanderTree,show.node.label=TRUE)
}
\keyword{datasets}
