\name{Unpub2}
\alias{Unpub2}
\docType{data}
\title{
Unpublished tree of fish
}
\description{
This has a single timetree of fish, made by Mike Alfaro and colleagues. It is not yet published.
}
\usage{data(Unpub2)}
\details{
Tree is unpublished, contact Mike Alfaro for more information}
\source{

}

\examples{
#not run
#data(Unpub2)


#best tree in ape (phylo) format:
#library(ape)
#fishTree<-Unpub2
#plot the tree
#plot(fishTree,show.node.label=TRUE)

}
\keyword{datasets}
