\name{Unpub3}
\alias{Unpub3}
\docType{data}
\title{
Giant glomogram of vertebrates. Use at your own risk.
}
\description{
This has a single timetree of ~15,000 vertebrates (Unpub3_tree). The tree is in phylo (\pkg{ape}) format.
}
\usage{data(Unpub3)}
\details{
Tree is unpublished, contact Jon Eastman for more information}
\source{

}

\examples{
#not run
#data(Unpub3)
#mTree<-Unpub3

#plot the tree
#plot(mTree,show.node.label=TRUE)
}
\keyword{datasets}
