\name{Drosophila2012}
\alias{Drosophila2012}
\alias{Drosophila2012_large}
\alias{Drosophila2012_small}
\docType{data}
\title{
Morales-Hojas and Vieira phylogeny of Drosophila
}
\description{
This has three trees made by BEAST for Drosophila, in two different files. Drosophila2012_large has 218 ingroup species and 9 calibration points (Drosophila2012_large[[1]]) or 5 calibration points (Drosophila2012_large[[2]]); Drosophila2012_small has 153 species (those with two or three genes) and 9 calibration points. Note that the D_ prefix in the original taxon names has been changed to Drosophila_
}
\usage{data(Drosophila2012_large)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Morales-Hojas R, Vieira J. 2012. Phylogenetic patterns of geographical and ecological diversification in the subgenus Drosophila. PLoS ONE 7(11): e49552.
}

\examples{
#not run
#data(Drosophila2012_large)
#plot(Drosophila2012_large[[1]],show.node.label=TRUE)
}
\keyword{datasets}
