\name{LohmannEtAl2013}
\alias{LohmannEtAl2013}
\docType{data}
\title{
Lohmann et al. 2013 tree of Bignonieae
}
\description{
This has a single chronogram (LohmannEtAl2013) from this paper. The tree is in phylo format.
}
\usage{data(LohmannEtAl2013)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Lohmann L.G., Bell C., Calio M., & Winkworth R.C. 2013. Pattern and timing of biogeographic history in the Neotropical tribe Bignonieae (Bignoniaceae). Botanical Journal of the Linnean Society, 171(1): 154-170.
}

\examples{
#not run
#data(LohmannEtAl2013)
#plot(LohmannEtAl2013,show.node.label=TRUE)
}
\keyword{datasets}
